forward msg fix
